+++ src/main/java/core/basesyntax/db/dao/StorageDaoImpl.java
@@ -5,12 +5,12 @@ import java.util.Map;
 
 public class StorageDaoImpl implements StorageDao {
     @Override
-    public void add(String fruitName, Long count) {
+    public void add(String fruitName, Integer count) {
         Storage.fruitsCount.put(fruitName, count);
     }
 
     @Override
-    public Map<String, Long> getInfo() {
+    public Map<String, Integer> getInfo() {
         return Storage.fruitsCount;
     }
 }
